chore: refactor, extract logic into custom hooks
