+++ dashboard/src/features/database/common/hooks/useIsDatabaseMigrating/index.ts
@@ -1,0 +1,1 @@
+export { default as useIsDatabaseMigrating } from './useIsDatabaseMigrating';
